Fixes #33484 - add an opt-in setting for the new host details page
