+++ webpack/assets/javascripts/react_app/components/HostDetails/consts.js
@@ -1,3 +1,6 @@
 export const DEFAULT_TAB = 'Overview';
 export const HOST_DETAILS_KEY = 'HOST_DETAILS';
 export const HOST_DETAILS_API_OPTIONS = { key: HOST_DETAILS_KEY };
+export const LEGACY_DETAILS_PATH = '/legacy/hosts'
+export const EXPERIMENTAL_HOST_DETAILS = '/experimental/hosts'
+export const TAB_PREFIX = '#/';
