+++ webpack/assets/javascripts/react_app/routes/ForemanSwitcher/OverrideRoutes.js
@@ -1,0 +1,1 @@
+export const OVERRIDE_ROUTES = [{ path: '/hosts/new', exact: true }];
