+++ webpack/assets/javascripts/react_app/routes/ForemanSwitcher/index.js
@@ -5,12 +5,16 @@ import { Switch, Route } from 'react-router-dom';
 
 import { fallbackRoute } from '../RoutingService';
 import { selectRoutes } from '../RouterSelector';
+import { OVERRIDE_ROUTES } from './OverrideRoutes';
 
 const ForemanSwitcher = ({ children: coreRoutes }) => {
   const routes = useSelector(() => selectRoutes(coreRoutes), shallowEqual);
 
   return (
     <Switch>
+      {OVERRIDE_ROUTES.map(({ path, exact }) => (
+        <Route path={path} exact={exact} render={fallbackRoute} key={path} />
+      ))}
       {routes}
       <Route render={fallbackRoute} key="default-route" />
     </Switch>
