+++ webpack/assets/javascripts/react_app/routes/HostDetails/constants.js
@@ -1,1 +1,2 @@
-export const HOST_DETAILS_PATH = '/experimental/hosts/:id';
+export const HOST_DETAILS_PATH = '/hosts/:id';
+export const HOST_DETAILS_EXPERIMENTAL_PATH = '/experimental/hosts/:id';
