+++ webpack/assets/javascripts/react_app/routes/HostDetails/experimental.js
@@ -1,0 +1,8 @@
+import React from 'react';
+import HostDetailsPage from '../../components/HostDetails/HostDetailsPage';
+import { HOST_DETAILS_EXPERIMENTAL_PATH } from './constants';
+
+export default {
+  path: HOST_DETAILS_EXPERIMENTAL_PATH,
+  render: props => <HostDetailsPage {...props} />,
+};
