+++ webpack/assets/javascripts/react_app/routes/HostDetails/index.js
@@ -4,5 +4,6 @@ import { HOST_DETAILS_PATH } from './constants';
 
 export default {
   path: HOST_DETAILS_PATH,
+  exact: true,
   render: props => <HostDetails {...props} />,
 };
